Split region loading and object population to factory class
