+++ 2006Redone Server/.gitignore
@@ -4,3 +4,4 @@
 .classpath
 .project
 settings.gradle
+/build/
